requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0
sse-starlette>=1.6.0
